refactor: перенести логику из FriendService в UserService
