+++ src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
@@ -16,6 +16,7 @@ import javax.validation.Valid;
 import java.util.*;
 import java.util.function.Function;
 
+import static java.util.Optional.ofNullable;
 import static java.util.stream.Collectors.*;
 
 @Service
@@ -25,13 +26,11 @@ public class UserService {
 
     private final AbstractRepository<Long, User> repository;
     private final Map<String, UserMapper<?, ?>> mapper;
-    private final FriendService friendService;
 
     @Autowired
-    public UserService(AbstractRepository<Long, User> repository, List<UserMapper<?, ?>> mappers, FriendService friendService) {
+    public UserService(AbstractRepository<Long, User> repository, List<UserMapper<?, ?>> mappers) {
         this.repository = repository;
         this.mapper = mappers.stream().collect(toMap(UserMapper::getKey, Function.identity()));
-        this.friendService = friendService;
     }
 
     public List<UserReadDto> findAll() {
@@ -77,9 +76,73 @@ public class UserService {
         if (maybeUser.isPresent()) {
             User user = maybeUser.get();
             for (Long friendId : user.getFriends()) {
-                repository.findById(friendId).map(friend -> friendService.removeFromFriends(id, friendId));
+                repository.findById(friendId).map(friend -> removeFromFriends(id, friendId));
             }
             return repository.delete(id);
+        }
+        return false;
+    }
+
+    public Optional<UserReadDto> addToFriends(Long userId, Long friendId) {
+        UserReadMapper userReadMapper = (UserReadMapper) mapper.get(UserReadMapper.class.getName());
+        var maybeUser = repository.findById(userId);
+        var maybeFriend = repository.findById(friendId);
+        UserReadDto userReadDto = null;
+        if (maybeUser.isPresent() && maybeFriend.isPresent()) {
+            User user = maybeUser.get();
+            User friend = maybeFriend.get();
+            user.getFriends().add(friend.getId());
+            friend.getFriends().add(user.getId());
+            repository.update(user);
+            repository.update(friend);
+            userReadDto = userReadMapper.mapFrom(user);
+        }
+        return ofNullable(userReadDto);
+    }
+
+    public Optional<List<UserReadDto>> findAllFriends(Long userId) {
+        UserReadMapper userReadMapper = (UserReadMapper) mapper.get(UserReadMapper.class.getName());
+        return repository.findById(userId)
+                .map(user -> user
+                        .getFriends()
+                        .stream()
+                        .map(repository::findById)
+                        .filter(Optional::isPresent)
+                        .map(friend -> userReadMapper.mapFrom(friend.get()))
+                        .collect(toList())
+                );
+    }
+
+    public Optional<List<UserReadDto>> findAllCommonFriends(Long userId, Long otherUserId) {
+        UserReadMapper userReadMapper = (UserReadMapper) mapper.get(UserReadMapper.class.getName());
+        List<UserReadDto> commonFriends = null;
+        var maybeUser = repository.findById(userId);
+        var maybeOtherUser = repository.findById(otherUserId);
+        if (maybeUser.isPresent() && maybeOtherUser.isPresent()) {
+            Set<Long> userFriends = new HashSet<>(maybeUser.get().getFriends());
+            Set<Long> otherUserFriends = new HashSet<>(maybeOtherUser.get().getFriends());
+            userFriends.retainAll(otherUserFriends);
+            commonFriends = userFriends
+                    .stream()
+                    .map(repository::findById)
+                    .filter(Optional::isPresent)
+                    .map(commonFriend -> userReadMapper.mapFrom(commonFriend.get()))
+                    .collect(toList());
+        }
+        return ofNullable(commonFriends);
+    }
+
+    public boolean removeFromFriends(Long userId, Long friendId) {
+        var maybeUser = repository.findById(userId);
+        var maybeFriend = repository.findById(friendId);
+        if (maybeUser.isPresent() && maybeFriend.isPresent()) {
+            User user = maybeUser.get();
+            User friend = maybeFriend.get();
+            user.getFriends().remove(friendId);
+            friend.getFriends().remove(userId);
+            repository.update(user);
+            repository.update(friend);
+            return true;
         }
         return false;
     }
